requirements.txt
3to2==1.1.1
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
autopep8==1.5.4
BASC-py4chan==0.6.5
beautifulsoup4==4.9.1
catalogue==2.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
cymem==2.0.3
Cython==0.29.21
decorator==4.4.2
deeppyer==1.0.1
discord==1.0.1
discord.py==1.5.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
html-text==0.5.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
mathparse==0.1.2
multidict==4.7.6
murmurhash==1.0.2
newsapi-python==0.2.6
nltk==3.5
numpy==1.19.2
opencv-python==4.4.0.46
packaging==20.4
pafy==0.5.5
peewee==3.13.3
Pillow==7.2.0
Pint==0.16
pipenv==2020.8.13
pokebase==1.3.0
praw==7.1.0
prawcore==1.5.0
preshed==3.0.2
psycopg2==2.8.5
Pybooru==4.2.2
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pymongo==3.11.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.7.5
python-decouple==3.3
python-ninegag==1.0
python-telegram-bot==12.8
pytz==2020.1
pyxlsb==1.0.6
PyYAML==3.13
regex==2020.7.14
requests==2.24.0
shorten-url==1.0.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.2.19
sqlparse==0.3.1
srsly==2.2.0
thinc==8.0.0a1
toml==0.10.1
tornado==6.0.4
tqdm==4.49.0
udpy==2.0.0
update-checker==0.18.0
urbandictionary==1.1
urllib3==1.25.9
virtualenv==20.0.31
virtualenv-clone==0.5.4
wasabi==0.8.0
websocket-client==0.57.0
Werkzeug==1.0.1
wikipedia==1.4.0
yarl==1.6.0
youtube-dl==2020.11.12
zdict==3.7.0
